update_firmware.sh: Force IPv4 connections
uclient-fetch simply returns with error if it happens to first try with
a IPv6 address and it can't (maybe because the device doesn't have IPv6
connectivity at all).
While this probably breaks IPv6-only connections, it's unclear we have
any of those yet and it fixes an actual problem we have with devices in
the wild.